Add tests for assign
